preproc: don't clear while expanding a mmacro.
This commit modifies the handling of the PP_CLEAR directive so that it doesn't
work if we are in the middle of a multi-line macro expansion.  This fixes a
use-after-free issue in the following situation:

```asm
%macro bar 0
%clear
%endmacro
bar
```

The fix just adds a check that `istk->expansion` doesn't contain any non-NULL
mmacro pointer, which is similar to the way it is checked in the `PP_UNMACRO`
directive.

Furthermore, this also fixes two other related issues:

* The token was not being advanced while checking for the %clear arguments,
  causing it to loop forever.

* There was not a goto done if one of the parameters to clear was not
  recognized, causing a null pointer dereference later on.